2.CameraBooth-main/CameraBooth/pythonCamera/requirements.txt
absl-py==2.1.0
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cvzone==1.6.1
cycler==0.12.1
flatbuffers==24.12.23
fonttools==4.55.3
idna==3.10
jax==0.4.38
jaxlib==0.4.38
kiwisolver==1.4.8
matplotlib==3.10.0
mediapipe==0.10.20
ml_dtypes==0.5.1
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
pillow==11.1.0
protobuf==4.25.5
pycparser==2.22
pyparsing==3.2.1
python-dateutil==2.9.0.post0
requests==2.32.3
scipy==1.15.1
segno==1.6.1
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
sxcu==4.1.0
urllib3==2.3.0
